Add getRtmpdumpCommand to Media, remove from Video - since doing that is wrong :( Also, make it work - omit the initial slash from the --app part.
